refactor(gateway): extract agent-cache/proxy/generation mixins from run.py (shard s5)
